Move ComponentList connect
