+++ ui/src/components/editor/ComponentList/index.js
@@ -1,3 +1,1 @@
-import { ComponentList } from './connect';
-
-export { ComponentList };
+export { ComponentList } from './ComponentList';
