+++ ui/src/components/index.js
@@ -7,3 +7,4 @@ export { Editor } from './editor/Editor';
 export { PlateTabBar } from './editor/PlateTabBar';
 export { Plate } from './editor/Plate';
 export { NoPlatesMessage } from './editor/Plate';
+export { ComponentList } from './editor/ComponentList';
