+++ ui/src/implementations/kapture-backend/components/index.js
@@ -7,3 +7,4 @@ export { Print } from './PrintConnected';
 export { Editor } from './EditorConnected';
 export { PlateTabBar } from './PlateTabBarConnected';
 export { Plate } from './PlateConnected';
+export { ComponentList } from './ComponentListConnected';
